# apps/cloud-backend/fusion-core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.AccountInfo:
    properties:
      description:
        example: Leading technology company
        type: string
      id:
        example: acc_987654321
        type: string
      name:
        example: Acme Corporation
        type: string
      type:
        example: Enterprise
        type: string
    type: object
...
  github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.CreateUserRequest:
    properties:
      account_id:
        example: acc_987654321
        type: string
      account_type_role_id:
        example: 2
        type: integer
      email:
        example: jane.smith@company.com
        type: string
      full_name:
        example: Jane Smith
        type: string
    required:
    - account_id
    - account_type_role_id
    - email
    - full_name
    type: object
...
  types.ErrorResponse:
    properties:
      error:
        example: Error message
        type: string
    type: object
...
  github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.RoleInfo:
    properties:
      id:
        example: 1
        type: integer
      role_name:
        example: Admin
        type: string
    type: object
...
  types.StatusOkForCreateUserProfile:
    properties:
      id:
        example: 53437319-7a5b-4462-bc7c-9e7f9a057a1a
        type: string
    type: object
  types.StatusOkForCreateUserSettings:
    properties:
      id:
        example: 53437319-7a5b-4462-bc7c-9e7f9a057a1a
        type: string
    type: object
  github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.UpdateUserRequest:
    properties:
      account_id:
        example: acc_111222333
        type: string
      account_type_role_id:
        example: 3
        type: integer
      full_name:
        example: John Smith
        type: string
    type: object
  types.UpdateUserSettingsRequest:
    properties:
      language:
        type: string
      theme:
        type: string
    type: object
...
  github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.UserAuthorizationResponse:
    properties:
      account:
        $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.AccountInfo'
      permissions:
        additionalProperties:
          type: string
        type: object
      role:
        $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.RoleInfo'
      user:
        $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.UserInfo'
    type: object
...
  github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.UserInfo:
    properties:
      email:
        example: john.doe@company.com
        type: string
      id:
        example: usr_123456789
        type: string
    type: object
  types.UserProfile:
    properties:
      address_line_1:
        type: string
      city:
        type: string
      client_type:
        type: string
      company_name:
        type: string
      company_website:
        type: string
      country:
        type: string
      created_at:
        type: string
      currency:
        type: string
      customer_type:
        type: string
      email:
        type: string
      first_name:
        type: string
      gdpr_opt_out:
        type: boolean
      id:
        type: string
      job_title:
        type: string
      last_name:
        type: string
      linked_profiles:
        type: object
      netsuite_customer_id:
        type: string
      phone:
        type: string
      price_list:
        type: object
      privacy_policy_accepted:
        type: boolean
      profile_photo_url:
        type: string
      state_province:
        type: string
      timezone:
        type: string
      unit_system:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      zip_postal_code:
        type: string
    type: object
  types.UserProfileUpdateRequest:
    properties:
      address_line_1:
        type: string
      city:
        type: string
      client_type:
        type: string
      company_name:
        type: string
      company_website:
        type: string
      country:
        type: string
      currency:
        type: string
      customer_type:
        type: string
      email:
        type: string
      first_name:
        type: string
      gdpr_opt_out:
        type: boolean
      job_title:
        type: string
      last_name:
        type: string
      linked_profiles:
        type: object
      netsuite_customer_id:
        type: string
      phone:
        type: string
      price_list:
        type: object
      privacy_policy_accepted:
        type: boolean
      profile_photo_url:
        type: string
      state_province:
        type: string
      timezone:
        type: string
      unit_system:
        type: string
      zip_postal_code:
        type: string
    type: object
  types.UserSettings:
    properties:
      created_at:
        type: string
      id:
        type: string
      language:
        type: string
      theme:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Fusion Cloud Backend API server providing comprehensive
    role-based access control, user management, and project management capabilities.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Fusion Cloud Backend API
  version: "1.0"
paths:
...
  /users:
    post:
      consumes:
      - application/json
      description: Create a new user with the specified details
      parameters:
      - description: User creation details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created user
          schema:
            $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.User'
        "400":
          description: Bad request - Invalid JSON payload
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - users
  /users/{email}:
    get:
      consumes:
      - application/json
      description: Get a specific user by their email address
      parameters:
      - description: User email address
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user
          schema:
            $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.User'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by email
      tags:
      - users
  /users/{userID}:
    patch:
      consumes:
      - application/json
      description: Update an existing user's information by their ID
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Updated user details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated user
          schema:
            $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.User'
        "400":
          description: Bad request - Invalid JSON payload
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /users/authorization:
    get:
      consumes:
      - application/json
      description: Get authorization details including user info, account, role, and
        permissions for the current authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user authorization details
          schema:
            $ref: '#/definitions/github_com_BoseProfessional_fusion-monorepo_apps_cloud-backend_fusion-core_internal_api_types.UserAuthorizationResponse'
        "401":
          description: Unauthorized - User email not found in token
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: User not found in the system
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user authorization details
      tags:
      - auth
  /users/profile:
    get:
      consumes:
      - application/json
      description: Get profile for a specific user
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user profile
          schema:
            $ref: '#/definitions/types.UserProfile'
        "404":
          description: User profile not found / Invalid user ID format
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile by user ID
      tags:
      - users/profile
    post:
      consumes:
      - application/json
      description: Create a new user profile with the provided details
      parameters:
      - description: User profile data
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/types.UserProfile'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created user profile
          schema:
            $ref: '#/definitions/types.StatusOkForCreateUserProfile'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new user profile
      tags:
      - users/profile
  /users/profile/:profileID:
    put:
      consumes:
      - application/json
      description: Update an existing user profile with the provided details
      parameters:
      - description: Profile ID
        in: path
        name: profileID
        required: true
        type: string
      - description: User profile data
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/types.UserProfileUpdateRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Successfully updated user profile
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an existing user profile
      tags:
      - users/profile
  /users/settings:
    get:
      consumes:
      - application/json
      description: Get settings for a specific user
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user settings
          schema:
            $ref: '#/definitions/types.UserSettings'
        "404":
          description: User settings not found / Invalid user ID format
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user settings by user ID
      tags:
      - users/settings
    post:
      consumes:
      - application/json
      description: Create a new user settings with the provided details
      parameters:
      - description: User settings data
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/types.UserSettings'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created user settings
          schema:
            $ref: '#/definitions/types.StatusOkForCreateUserSettings'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Create a new user settings
      tags:
      - users/settings
  /users/settings/:settingsID:
    put:
      consumes:
      - application/json
      description: Update an existing user settings with the provided details
      parameters:
      - description: Settings ID
        in: path
        name: settingsID
        required: true
        type: string
      - description: User settings data
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/types.UpdateUserSettingsRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Successfully updated user settings
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: User settings not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an existing user settings
      tags:
      - users/settings
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"